+++ packages/vessel/src/configs/eslint.ts
@@ -1,12 +1,24 @@
 import { ESLintConfig } from '@beemo/driver-eslint';
-import { IGNORE_PATHS } from '@manifest-ui/vessel-constants';
 
 const { tool } = process.beemo;
 const { react } = tool.config.settings;
 
 const config = {
-  extends: ['@manifest-ui/eslint-config'],
-  ignore: [...IGNORE_PATHS, '*.min.js', '*.map', '*.snap'],
+  extends: ['vessel'],
+  ignore: [
+    'build/',
+    'coverage/',
+    'node_modules/',
+    'public/',
+    'esm/',
+    'lib/',
+    'tmp/',
+    'dist/',
+    '*.d.ts',
+    '*.min.js',
+    '*.map',
+    '*.snap',
+  ],
   root: true,
 };