+++ packages/vessel/CHANGELOG.md
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 # @manifest-ui/vessel
 
 ## 0.1.0
@@ -7,10 +8,22 @@
 - [#11](https://github.com/project44/vessel/pull/11)
   [`9d75241`](https://github.com/project44/vessel/commit/9d752412a1ee1adfb52c6393b4323ef969ca5dfe)
   Thanks [@branclar](https://github.com/branclar)! - Change package names
+=======
+# @manifest-ui/vessel-dev
+
+## 0.2.0
+
+### Minor Changes
+
+- [#8](https://github.com/project44/vessel/pull/8)
+  [`aec7518`](https://github.com/project44/vessel/commit/aec75181429d6ab16d08a34ee37b184bde49cd24)
+  Thanks [@bclark-p44](https://github.com/bclark-p44)! - Repo re-organization
+>>>>>>> d680b9f (fix: fixing packages)
 
 ### Patch Changes
 
 - Updated dependencies
+<<<<<<< HEAD
   [[`9d75241`](https://github.com/project44/vessel/commit/9d752412a1ee1adfb52c6393b4323ef969ca5dfe)]:
   - @manifest-ui/babel-preset@0.1.0
   - @manifest-ui/eslint-config@0.1.0
@@ -18,3 +31,12 @@
   - @manifest-ui/prettier-config@0.1.0
   - @manifest-ui/tsconfig@0.1.0
   - @manifest-ui/vessel-constants@0.1.2
+=======
+  [[`aec7518`](https://github.com/project44/vessel/commit/aec75181429d6ab16d08a34ee37b184bde49cd24)]:
+  - babel-preset-vessel@0.1.1
+  - @manifest-ui/vessel-constants@0.1.1
+  - eslint-config-vessel@0.1.1
+  - jest-preset-vessel@0.1.1
+  - prettier-config-vessel@0.1.1
+  - tsconfig-vessel@0.1.1
+>>>>>>> d680b9f (fix: fixing packages)